code updated

+++ src/app/customer/pages/customer-group/customer-group.component.ts
@@ -5,7 +5,7 @@ import { Subject } from 'rxjs';
 import { takeUntil, startWith, debounceTime, distinctUntilChanged, switchMap } from 'rxjs/operators';
 import { CustomerService } from '../../../shared/services/customer.service';
 import { UtilityService } from '../../../shared/utility/utility.service';
-import { LazyLoadEvent } from 'primeng/api';
+import { LazyLoadEvent, ConfirmationService } from 'primeng/api';
 
 interface Action {
   name: string,
@@ -27,6 +27,7 @@ export class CustomerGroupComponent implements OnInit {
   page: number = 0;
   searchKey: string;
   searchValue: string;
+  action:string;
   Date = new Date();
 
   @ViewChild(Table) tableComponent: Table;
@@ -41,7 +42,8 @@ export class CustomerGroupComponent implements OnInit {
     private router: Router, 
     private route : ActivatedRoute,
     private customerService: CustomerService,
-    private utilityService: UtilityService
+    private utilityService: UtilityService,
+    private confirmationService: ConfirmationService
     ) {}
 
   ngOnInit() {
@@ -103,8 +105,8 @@ export class CustomerGroupComponent implements OnInit {
   }
 
 
-  getAllCustomers() {
-    this.customerService.getCustomerGroup().subscribe(
+  getAllCustomers(page) {
+    this.customerService.getCustomerGroup(page).subscribe(
       (success: any) => {
         this.customerList = success.data.results;
         this.totalCount = success.data.total;
@@ -144,11 +146,16 @@ export class CustomerGroupComponent implements OnInit {
     this.page = event.first / 10;
     // if there is a search term present in the search bar, then paginate with the search term
     if (!this.searchBar) {
-      this.getAllCustomers();
+      this.getAllCustomers(this.page);
+      this.utilityService.loaderStop();
+
     } else {
       this.getAllCustomersSearch(this.page, this.searchBar);
+      this.utilityService.loaderStop();
+
     }
   }
+
 
   onAddCustomerGroup(){
     this.router.navigate(['../newGroup'],{relativeTo : this.route})
@@ -160,18 +167,26 @@ export class CustomerGroupComponent implements OnInit {
     if(event.value === 'Delete') {
 
       console.log('delete id', id);
-      this.customerService.deleteCustomerGroup(id).pipe(takeUntil(this._unsubscribe)).subscribe(
-        (success: any) => {
-          console.log(success);
-          this.customerList = this.customerList.filter((item: any) => {
-            return id !== item.id
-          })
-          // this.initiateSearch();
+      this.confirmationService.confirm({
+        message: 'Are you sure that you want to perform this action?',
+        accept: () => {
+          this.customerService.deleteCustomerGroup(id).pipe(takeUntil(this._unsubscribe)).subscribe(
+            (success: any) => {
+              console.log(success);
+              this.getAllCustomers(this.page);
+              // this.initiateSearch();
+            },
+            error => {
+              console.log(error);
+            }
+          )
         },
-        error => {
-          console.log(error);
+        reject: () => {
+          this.action = null;
         }
-      )
+      
+      });
+
     }
   }
